+++ tests/testthat/helper-utils.R
@@ -21,3 +21,26 @@ skip_pandoc <- function() {
   has_pandoc <- rmarkdown::pandoc_available(version = "1.12.3", error = FALSE)
   skip_if_not(has_pandoc, "no pandoc >= 1.12.3")
 }
+
+targets_1.3.2.9004 <- function() {
+  utils::compareVersion(
+    a = as.character(packageVersion("targets")),
+    b = "1.3.2.9004"
+  ) > -1L
+}
+
+status_completed <- function() {
+  if (targets_1.3.2.9004()) {
+    "completed"
+  } else {
+    "built"
+  }
+}
+
+status_dispatched <- function() {
+  if (targets_1.3.2.9004()) {
+    "dispatched"
+  } else {
+    "started"
+  }
+}