+++ tests/testthat/test-tar_hook_before.R
@@ -162,7 +162,7 @@ targets::tar_test("hook invalidates target", {
   targets::tar_make(callr_function = NULL)
   out <- targets::tar_progress()
   expect_equal(out$name, "a")
-  expect_equal(out$progress, "built")
+  expect_equal(out$progress, status_completed())
 })
 
 targets::tar_test("tar_hook_before() sets deps by default", {