+++ tests/testthat/test-tar_quarto_rep.R
@@ -184,7 +184,7 @@ targets::tar_test("tar_quarto_rep() run", {
   out <- out[grepl("^report", out$name), ]
   expect_equal(nrow(out), 3)
   out <- out$progress[!(out$name %in% c("report", "report_params"))]
-  expect_equal(out, "built")
+  expect_equal(out, status_completed())
 })
 
 targets::tar_test("tar_quarto_rep() run with output_file specified", {